+++ benches/benches/shuffle.rs
@@ -35,10 +35,10 @@ pub fn bench(c: &mut Criterion) {
     bench_rng::<rand_pcg::Pcg64>(c, "Pcg64");
 }
 
-fn bench_rng<Rng: RngCore + SeedableRng>(c: &mut Criterion, rng_name: &'static str) {
+fn bench_rng<R: Rng + SeedableRng>(c: &mut Criterion, rng_name: &'static str) {
     for length in [1, 2, 3, 10, 100, 1000, 10000].map(black_box) {
         c.bench_function(format!("shuffle_{length}_{rng_name}").as_str(), |b| {
-            let mut rng = Rng::seed_from_u64(123);
+            let mut rng = R::seed_from_u64(123);
             let mut vec: Vec<usize> = (0..length).collect();
             b.iter(|| {
                 vec.shuffle(&mut rng);
@@ -49,7 +49,7 @@ fn bench_rng<Rng: RngCore + SeedableRng>(c: &mut Criterion, rng_name: &'static s
         if length >= 10 {
             let name = format!("partial_shuffle_{length}_{rng_name}");
             c.bench_function(name.as_str(), |b| {
-                let mut rng = Rng::seed_from_u64(123);
+                let mut rng = R::seed_from_u64(123);
                 let mut vec: Vec<usize> = (0..length).collect();
                 b.iter(|| {
                     vec.partial_shuffle(&mut rng, length / 2);