+++ src/rng.rs
@@ -7,21 +7,21 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-//! [`Rng`] trait
+//! [`RngExt`] trait
 
 use crate::distr::uniform::{SampleRange, SampleUniform};
 use crate::distr::{self, Distribution, StandardUniform};
 use core::num::Wrapping;
 use core::{mem, slice};
-use rand_core::RngCore;
+use rand_core::Rng;
 
 /// User-level interface for RNGs
 ///
-/// [`RngCore`] is the `dyn`-safe implementation-level interface for Random
+/// [`Rng`] is the `dyn`-safe implementation-level interface for Random
 /// (Number) Generators. This trait, `Rng`, provides a user-level interface on
-/// RNGs. It is implemented automatically for any `R: RngCore`.
+/// RNGs. It is implemented automatically for any `R: Rng`.
 ///
-/// This trait must usually be brought into scope via `use rand::Rng;` or
+/// This trait must usually be brought into scope via `use rand::RngExt;` or
 /// `use rand::prelude::*;`.
 ///
 /// # Generic usage
@@ -29,11 +29,9 @@ use rand_core::RngCore;
 /// The basic pattern is `fn foo<R: Rng + ?Sized>(rng: &mut R)`. Some
 /// things are worth noting here:
 ///
-/// - Since `Rng: RngCore` and every `RngCore` implements `Rng`, it makes no
-///   difference whether we use `R: Rng` or `R: RngCore`.
-/// - The `+ ?Sized` un-bounding allows functions to be called directly on
-///   type-erased references; i.e. `foo(r)` where `r: &mut dyn RngCore`. Without
-///   this it would be necessary to write `foo(&mut r)`.
+/// - Since `RngExt: Rng` and every `RngExt` implements `Rng`, it makes no
+///   difference whether we use `R: Rng` or `R: RngExt` for `R: Sized`.
+/// - Only `Rng` is dyn safe, supporting `&mut dyn Rng` and `R: Rng + ?Sized`.
 ///
 /// An alternative pattern is possible: `fn foo<R: Rng>(rng: R)`. This has some
 /// trade-offs. It allows the argument to be consumed directly without a `&mut`
@@ -47,7 +45,7 @@ use rand_core::RngCore;
 /// Example:
 ///
 /// ```
-/// use rand::Rng;
+/// use rand::{Rng, RngExt};
 ///
 /// fn foo<R: Rng + ?Sized>(rng: &mut R) -> f32 {
 ///     rng.random()
@@ -55,13 +53,13 @@ use rand_core::RngCore;
 ///
 /// # let v = foo(&mut rand::rng());
 /// ```
-pub trait Rng: RngCore {
+pub trait RngExt: Rng {
     /// Return a random value via the [`StandardUniform`] distribution.
     ///
     /// # Example
     ///
     /// ```
-    /// use rand::Rng;
+    /// use rand::RngExt;
     ///
     /// let mut rng = rand::rng();
     /// let x: u32 = rng.random();
@@ -76,11 +74,11 @@ pub trait Rng: RngCore {
     /// generated.
     ///
     /// For arrays of integers, especially for those with small element types
-    /// (< 64 bit), it will likely be faster to instead use [`Rng::fill`],
+    /// (< 64 bit), it will likely be faster to instead use [`RngExt::fill`],
     /// though note that generated values will differ.
     ///
     /// ```
-    /// use rand::Rng;
+    /// use rand::RngExt;
     ///
     /// let mut rng = rand::rng();
     /// let tuple: (u8, i32, char) = rng.random(); // arbitrary tuple support
@@ -101,7 +99,7 @@ pub trait Rng: RngCore {
 
     /// Return an iterator over [`random`](Self::random) variates
     ///
-    /// This is a just a wrapper over [`Rng::sample_iter`] using
+    /// This is a just a wrapper over [`RngExt::sample_iter`] using
     /// [`distr::StandardUniform`].
     ///
     /// Note: this method consumes its argument. Use
@@ -110,7 +108,7 @@ pub trait Rng: RngCore {
     /// # Example
     ///
     /// ```
-    /// use rand::{rngs::SmallRng, Rng, SeedableRng};
+    /// use rand::{rngs::SmallRng, RngExt, SeedableRng};
     ///
     /// let rng = SmallRng::seed_from_u64(0);
     /// let v: Vec<i32> = rng.random_iter().take(5).collect();
@@ -141,7 +139,7 @@ pub trait Rng: RngCore {
     /// # Example
     ///
     /// ```
-    /// use rand::Rng;
+    /// use rand::RngExt;
     ///
     /// let mut rng = rand::rng();
     ///
@@ -175,7 +173,7 @@ pub trait Rng: RngCore {
     /// # Example
     ///
     /// ```
-    /// use rand::Rng;
+    /// use rand::RngExt;
     ///
     /// let mut rng = rand::rng();
     /// println!("{}", rng.random_bool(1.0 / 3.0));
@@ -213,7 +211,7 @@ pub trait Rng: RngCore {
     /// # Example
     ///
     /// ```
-    /// use rand::Rng;
+    /// use rand::RngExt;
     ///
     /// let mut rng = rand::rng();
     /// println!("{}", rng.random_ratio(2, 3));
@@ -237,7 +235,7 @@ pub trait Rng: RngCore {
     /// ### Example
     ///
     /// ```
-    /// use rand::Rng;
+    /// use rand::RngExt;
     /// use rand::distr::Uniform;
     ///
     /// let mut rng = rand::rng();
@@ -258,7 +256,7 @@ pub trait Rng: RngCore {
     /// # Example
     ///
     /// ```
-    /// use rand::Rng;
+    /// use rand::RngExt;
     /// use rand::distr::{Alphanumeric, Uniform, StandardUniform};
     ///
     /// let mut rng = rand::rng();
@@ -295,7 +293,7 @@ pub trait Rng: RngCore {
     ///
     /// This method is implemented for types which may be safely reinterpreted
     /// as an (aligned) `[u8]` slice then filled with random data. It is often
-    /// faster than using [`Rng::random`] but not value-equivalent.
+    /// faster than using [`RngExt::random`] but not value-equivalent.
     ///
     /// The distribution is expected to be uniform with portable results, but
     /// this cannot be guaranteed for third-party implementations.
@@ -303,19 +301,19 @@ pub trait Rng: RngCore {
     /// # Example
     ///
     /// ```
-    /// use rand::Rng;
+    /// use rand::RngExt;
     ///
     /// let mut arr = [0i8; 20];
     /// rand::rng().fill(&mut arr[..]);
     /// ```
     ///
-    /// [`fill_bytes`]: RngCore::fill_bytes
+    /// [`fill_bytes`]: Rng::fill_bytes
     #[track_caller]
     fn fill<T: Fill>(&mut self, dest: &mut [T]) {
         Fill::fill_slice(dest, self)
     }
 
-    /// Alias for [`Rng::random`].
+    /// Alias for [`RngExt::random`].
     #[inline]
     #[deprecated(
         since = "0.9.0",
@@ -328,7 +326,7 @@ pub trait Rng: RngCore {
         self.random()
     }
 
-    /// Alias for [`Rng::random_range`].
+    /// Alias for [`RngExt::random_range`].
     #[inline]
     #[deprecated(since = "0.9.0", note = "Renamed to `random_range`")]
     fn gen_range<T, R>(&mut self, range: R) -> T
@@ -339,14 +337,14 @@ pub trait Rng: RngCore {
         self.random_range(range)
     }
 
-    /// Alias for [`Rng::random_bool`].
+    /// Alias for [`RngExt::random_bool`].
     #[inline]
     #[deprecated(since = "0.9.0", note = "Renamed to `random_bool`")]
     fn gen_bool(&mut self, p: f64) -> bool {
         self.random_bool(p)
     }
 
-    /// Alias for [`Rng::random_ratio`].
+    /// Alias for [`RngExt::random_ratio`].
     #[inline]
     #[deprecated(since = "0.9.0", note = "Renamed to `random_ratio`")]
     fn gen_ratio(&mut self, numerator: u32, denominator: u32) -> bool {
@@ -354,7 +352,7 @@ pub trait Rng: RngCore {
     }
 }
 
-impl<R: RngCore + ?Sized> Rng for R {}
+impl<R: Rng + ?Sized> RngExt for R {}
 
 /// Support filling a slice with random data
 ///
@@ -563,7 +561,7 @@ mod test {
 
     #[test]
     fn test_rng_mut_ref() {
-        fn use_rng(mut r: impl Rng) {
+        fn use_rng(mut r: impl RngExt) {
             let _ = r.next_u32();
         }
 
@@ -575,7 +573,7 @@ mod test {
     fn test_rng_trait_object() {
         use crate::distr::{Distribution, StandardUniform};
         let mut rng = rng(109);
-        let mut r = &mut rng as &mut dyn RngCore;
+        let mut r = &mut rng as &mut dyn Rng;
         r.next_u32();
         r.random::<i32>();
         assert_eq!(r.random_range(0..1), 0);
@@ -587,7 +585,7 @@ mod test {
     fn test_rng_boxed_trait() {
         use crate::distr::{Distribution, StandardUniform};
         let rng = rng(110);
-        let mut r = Box::new(rng) as Box<dyn RngCore>;
+        let mut r = Box::new(rng) as Box<dyn Rng>;
         r.next_u32();
         r.random::<i32>();
         assert_eq!(r.random_range(0..1), 0);