+++ src/rngs/xoshiro128plusplus.rs
@@ -7,7 +7,7 @@
 // except according to those terms.
 
 use core::convert::Infallible;
-use rand_core::{SeedableRng, TryRngCore, utils};
+use rand_core::{SeedableRng, TryRng, utils};
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
@@ -64,7 +64,7 @@ impl SeedableRng for Xoshiro128PlusPlus {
     }
 }
 
-impl TryRngCore for Xoshiro128PlusPlus {
+impl TryRng for Xoshiro128PlusPlus {
     type Error = Infallible;
 
     #[inline]
@@ -102,7 +102,7 @@ impl TryRngCore for Xoshiro128PlusPlus {
 #[cfg(test)]
 mod tests {
     use super::Xoshiro128PlusPlus;
-    use rand_core::{RngCore, SeedableRng};
+    use rand_core::{Rng, SeedableRng};
 
     #[test]
     fn reference() {