+++ src/seq/index.rs
@@ -13,9 +13,9 @@ use core::{hash::Hash, ops::AddAssign};
 // BTreeMap is not as fast in tests, but better than nothing.
 #[cfg(feature = "std")]
 use super::WeightError;
-use crate::Rng;
 use crate::distr::uniform::SampleUniform;
 use crate::distr::{Distribution, Uniform};
+use crate::{Rng, RngExt};
 #[cfg(not(feature = "std"))]
 use alloc::collections::BTreeSet;
 #[cfg(feature = "serde")]